refactor: emprolijando librerias
