+++ org.uqbar.project.wollok/META-INF/MANIFEST.MF
@@ -42,7 +42,7 @@ Export-Package: org.uqbar.project.wollok,
  org.uqbar.project.wollok.interpreter.natives,
  org.uqbar.project.wollok.interpreter.operation,
  org.uqbar.project.wollok.interpreter.stack,
- org.uqbar.project.wollok.manifest,
+ org.uqbar.project.wollok.libraries,
  org.uqbar.project.wollok.model,
  org.uqbar.project.wollok.nature,
  org.uqbar.project.wollok.parser.antlr,
