+++ org.uqbar.project.wollok/src/org/uqbar/project/wollok/WollokDslRuntimeModule.xtend
@@ -24,8 +24,8 @@ import org.uqbar.project.wollok.interpreter.natives.DefaultNativeObjectFactory
 import org.uqbar.project.wollok.interpreter.natives.NativeObjectFactory
 import org.uqbar.project.wollok.linking.WollokLinker
 import org.uqbar.project.wollok.linking.WollokLinkingDiagnosticMessageProvider
-import org.uqbar.project.wollok.manifest.JarWollokLibraries
-import org.uqbar.project.wollok.manifest.WollokLibraries
+import org.uqbar.project.wollok.libraries.JarWollokLibraries
+import org.uqbar.project.wollok.libraries.WollokLibraries
 import org.uqbar.project.wollok.parser.WollokSyntaxErrorMessageProvider
 import org.uqbar.project.wollok.scoping.WollokGlobalScopeProvider
 import org.uqbar.project.wollok.scoping.WollokImportedNamespaceAwareLocalScopeProvider
@@ -35,11 +35,11 @@ import org.uqbar.project.wollok.scoping.cache.MapBasedWollokGlobalScopeCache
 import org.uqbar.project.wollok.scoping.cache.WollokGlobalScopeCache
 import org.uqbar.project.wollok.serializer.WollokDslSyntacticSequencerWithSyntheticLinking
 import org.uqbar.project.wollok.utils.DummyJvmTypeProviderFactory
-import org.uqbar.project.wollok.manifest.WollokLibraryLoader
-import org.uqbar.project.wollok.manifest.BasicWollokLibraryLoader
+import org.uqbar.project.wollok.libraries.WollokLibraryLoader
+import org.uqbar.project.wollok.libraries.BasicWollokLibraryLoader
 import com.google.inject.TypeLiteral
-import org.uqbar.project.wollok.manifest.WollokLib
-import org.uqbar.project.wollok.manifest.StandardWollokLib
+import org.uqbar.project.wollok.libraries.WollokLib
+import org.uqbar.project.wollok.libraries.StandardWollokLib
 
 /**
  * Use this class to register components to be used at runtime / without the Equinox extension registry.
