+++ org.uqbar.project.wollok/src/org/uqbar/project/wollok/libraries/JarWollokLibraries.xtend
@@ -1,4 +1,4 @@
-package org.uqbar.project.wollok.manifest
+package org.uqbar.project.wollok.libraries
 
 import java.util.List
 import org.eclipse.emf.ecore.resource.Resource
@@ -6,6 +6,10 @@ import org.eclipse.xtext.resource.IResourceDescription.Manager
 import com.google.inject.Inject
 import com.google.inject.name.Named
 
+/**
+ * It finds all JarWollokLib related to a wollok file represented by resource
+ * @author leo
+ */
 class JarWollokLibraries implements WollokLibraries {
 	
 	@Inject
