+++ org.uqbar.project.wollok/src/org/uqbar/project/wollok/libraries/WollokLib.java
@@ -1,11 +1,11 @@
-package org.uqbar.project.wollok.manifest;
+package org.uqbar.project.wollok.libraries;
 
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.xtext.resource.IEObjectDescription;
 import org.eclipse.xtext.resource.IResourceDescription;
 
 /**
- * Represent a wollok library. 
+ * It represents a wollok library. 
  * A Wollok library is able to add objects and classes to a wollok project
  * @author leo
  */
