+++ org.uqbar.project.wollok/src/org/uqbar/project/wollok/libraries/WollokLibraries.java
@@ -1,9 +1,14 @@
-package org.uqbar.project.wollok.manifest;
+package org.uqbar.project.wollok.libraries;
 
 import java.util.List;
 
 import org.eclipse.emf.ecore.resource.Resource;
 
+/**
+ * It finds all external libraries related to a wollok file 
+ * represented by resource
+ * @author leo
+ */
 public interface WollokLibraries {
 	public List<WollokLib> getWollokLibs(Resource resource);
 }
